Allow using Accessory in normal unit tests
Moves the Accessory from test/e2e/framework into pkg/testhelper, so it
can be used in unit tests.
